requirements.txt
pandas
scikit-learn
streamlit
google-generativeai
requests
streamlit-calendar
python-multipart